+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+- Move away from Addressable [#1051](https://github.com/bridgetownrb/bridgetown/issues/1051) ([@anantone](https://github.com/anantone))
 .
 
 ## [2.1.0.beta1] - 2025-12-03